// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};
// ...
export type CentraldeDecorado_Centraldedecorados = AcfFieldGroup & {
  __typename?: 'CentraldeDecorado_Centraldedecorados';
  atendimentoDomingo?: Maybe<Scalars['String']>;
  atendimentoFeriado?: Maybe<Scalars['String']>;
  atendimentoSabado?: Maybe<Scalars['String']>;
  atendimentoSemana?: Maybe<Scalars['String']>;
  enderecoBairro?: Maybe<Scalars['String']>;
  enderecoCidade?: Maybe<Scalars['String']>;
  enderecoNumero?: Maybe<Scalars['String']>;
  enderecoRua?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  /** Resolução recomendada: 1920 x 1080px. */
  imagemPrincipal?: Maybe<MediaItem>;
  telefone?: Maybe<Scalars['String']>;
  whatsapp?: Maybe<Scalars['String']>;
};